code review fixes for null to undefined
